Seperating out the initialization and ingest processes. TODO: Need to set it up so that ingest does not necessarily ingest all of the data each and every single time; only should need to pull in newer data.
