+++ src/ingest.py
@@ -1,4 +1,4 @@
-# Initialize the tables for the DB.
+# Ingest the data into the DB.
 
 import argparse as ap
 import duckdb as db
@@ -7,41 +7,24 @@ import polars as pl
 from loguru import logger
 from statsapi import get
 
-def parse_args() -> str:
+def parse_args() -> tuple[str, str]:
     # Define the argument parser for the command line.
     parser = ap.ArgumentParser("Initialize the tables.")
     # Define an accepted argument.
     parser.add_argument("-p", "--path", help="Path to the DB file.")
+    parser.add_argument("-c", "--complete", help="y/n to ingest the all historical and current data.")
     # Pull the argument from the command line input.
     args = parser.parse_args()
     # Return the argument.
-    return args.path
-
+    return args.path, args.complete
 
 
 def seasons_ingest(con:db.DuckDBPyConnection) -> None:
-    """Initialize a table of the Seasons for the MLB.
+    """Ingest the data for the seasons for the MLB.
 
     Args:
         con (duckdb.duckDBCOnnection): The connection to the database.
     """
-    # Initialize the table.
-    con.execute(
-        """
-        create table seasons (
-            season_id varchar(4)
-            , has_wildcard boolean
-            , preseason_start date
-            , season_start date
-            , regular_season_start date
-            , regular_season_end date
-            , season_end date
-            , offseason_start date
-            , offSeason_end date
-            , primary key (season_id)
-        );
-        """
-    )
     # Insert the data into the table.
     # - Pull the season data from the api.
     seasons = get("seasons", params={"all":True, "sportId":1})
@@ -76,19 +59,6 @@ def teams_ingest(con:db.DuckDBPyConnection) -> None:
     Args:
         con (duckdb.DuckDBPyConnection): The connection to the database.
     """
-    # Initialize the table.
-    con.execute(
-        """
-        create table teams (
-            season_id varchar(4)
-            , team_id varchar(4)
-            , team_name varchar(50)
-            , team_abbr varchar(4)
-            , foreign key (season_id) references seasons(season_id)
-            , primary key (season_id, team_id)
-        );
-        """
-    )
     # Pull the data from the API.
     # 1. Get a list of seasons.
     seasons = con.sql(
@@ -135,29 +105,6 @@ def schedule_ingest(con:db.DuckDBPyConnection) -> None:
     Args:
         con (duckdb.duckDBConnection): The connection to the database.
     """
-    # Initialize the table.
-    con.execute(
-        """
-        create table schedule (
-            season_id varchar(4)
-            , game_date date
-            , game_id varchar(50)
-            , double_header varchar(1)
-            , away_team varchar(4)
-            , away_team_wins integer
-            , away_team_losses integer
-            , home_team varchar(4)
-            , home_team_wins integer
-            , home_team_losses integer
-            , foreign key (season_id) references seasons(season_id)
-            -- Would have these keys set, but the api data has 
-            -- duplicate/missing values for these...
-            -- , foreign key (season_id, away_team) references teams(season_id, team_id)
-            -- , foreign key (season_id, home_team) references teams(season_id, team_id)
-            -- , primary key (season_id, game_id)
-        )
-        """
-    )
     # Pull the data from the api.
     # 1. Get a list of seasons.
     seasons = con.sql(
@@ -227,10 +174,10 @@ def main():
     """Main function"""
     logger.info("Data ingestion beginning.")
     # Pull in the connection from the argument.
-    path = parse_args()
+    path, complete = parse_args()
     # Connect to the db.
     con = db.connect(path)
-    logger.success(f"Connected to the DB at: {path}.")
+    logger.info(f"Connected to the DB at: {path}.")
     # Season ingest.
     seasons_ingest(con)
     logger.success("Ingested data for each season.")
