--- table_init.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/bash
-
-poetry run python -m src.db_init -p ./db/twenty_five.db
